requirements.txt
django>=5.0a1
djangorestframework>=3.12
coverage
tox
pre-commit
